agents/research-agent/requirements.txt
# Research Agent V2 Requirements
# Production-ready dependencies with Langfuse observability

# Core dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Langfuse observability
langfuse>=2.0.0

# LangGraph for workflow orchestration
langgraph>=0.0.20
langchain>=0.1.0
langchain-community>=0.0.10

# Model providers
openai>=1.0.0
anthropic>=0.15.0
google-generativeai>=0.3.0
groq>=0.4.0
huggingface-hub>=0.19.0
transformers>=4.35.0
sentence-transformers>=2.2.0

# Search and research tools
duckduckgo-search>=4.0.0
arxiv>=2.0.0
wikipedia-api>=0.6.0
newsapi-python>=0.2.7
beautifulsoup4>=4.12.0
requests>=2.31.0
aiohttp>=3.9.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0

# NLP and analysis
nltk>=3.8.0
spacy>=3.7.0
textblob>=0.17.0
vaderSentiment>=3.3.2

# Citation and academic tools
bibtexparser>=1.4.0
scholarly>=1.7.0

# Fact checking and validation
fact-checker>=0.1.0

# Async support
asyncio>=3.4.3
aiofiles>=23.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Logging and monitoring
loguru>=0.7.0

# Performance
cachetools>=5.3.0
redis>=5.0.0

# Security
cryptography>=41.0.0

# Optional enhanced features
# Uncomment if needed:
# crewai>=0.1.0
# autogen>=0.2.0
# llamaindex>=0.9.0